AGIAN/requirements.txt
gensim==3.4.0
numpy==1.19.5
pandas==1.4.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.15.0
tensorflow-gpu==2.5.0
tokenizers==0.10.3
tqdm==4.64.1
transformers==4.6.0
spacy==3.2.3
